+++ frontend/catalyst/device/qjit_device.py
@@ -39,9 +39,13 @@ from catalyst.utils.patching import Patcher
 from catalyst.utils.runtime_environment import get_lib_path
 from catalyst.utils.toml import (
     DeviceCapabilities,
+    TOMLDocument,
+    ProgramFeatures,
     OperationProperties,
     intersect_operations,
     pennylane_operation_set,
+    read_toml_file,
+    load_device_capabilities,
 )
 
 RUNTIME_OPERATIONS = [
@@ -523,3 +527,47 @@ def validate_device_capabilities(
                 "Observables in qml.device.observables and specification file do not match.\n"
                 f"Observables that present only in spec: {spec_observables - device_observables}\n"
             )
+
+
+def get_device_toml_config(device) -> TOMLDocument:
+    """Get the contents of the device config file."""
+    if hasattr(device, "config"):
+        # The expected case: device specifies its own config.
+        toml_file = device.config
+    else:
+        # TODO: Remove this section when `qml.Device`s are guaranteed to have their own config file
+        # field.
+        device_lpath = pathlib.Path(get_lib_path("runtime", "RUNTIME_LIB_DIR"))
+
+        name = device.short_name if isinstance(device, qml.Device) else device.name
+        # The toml files name convention we follow is to replace
+        # the dots with underscores in the device short name.
+        toml_file_name = name.replace(".", "_") + ".toml"
+        # And they are currently saved in the following directory.
+        toml_file = device_lpath.parent / "lib" / "backend" / toml_file_name
+
+    try:
+        config = read_toml_file(toml_file)
+    except FileNotFoundError as e:
+        raise CompileError(
+            "Attempting to compile program for incompatible device: "
+            f"Config file ({toml_file}) does not exist"
+        ) from e
+
+    return config
+
+
+def get_device_capabilities(
+    device, program_features: Optional[ProgramFeatures] = None
+) -> DeviceCapabilities:
+    """Get or load DeviceCapabilities structure from device"""
+
+    if hasattr(device, "qjit_capabilities"):
+        return device.qjit_capabilities
+    else:
+        program_features = (
+            program_features if program_features else ProgramFeatures(device.shots is not None)
+        )
+        device_name = device.short_name if isinstance(device, qml.Device) else device.name
+        device_config = get_device_toml_config(device)
+        return load_device_capabilities(device_config, program_features, device_name)
